chore: Update CITATION.cff metadata
This commit updates the metadata in the CITATION.cff file. It fixes the cff-version to "FIXME" and updates the title to "DG_RoomAcoustics". It also updates the version to "2024.04" and the repository code to "https://github.com/Building-acoustics-TU-Eindhoven/edg-acoustics". Additionally, it adds new keywords related to wave-based finite element method and impulse response in room acoustics. The message in the file is also modified to request users to cite the software if they use it.