feat: Add CoeffByVF.m script for approximating reflection coefficient using vector fitting
This commit adds the CoeffByVF.m script, which approximates the reflection coefficient of a material using vector fitting. The script is used as input for the DG simulation and includes user-defined parameters for frequency range, number of iterations, and number of poles for the approximation. The resulting state space model is saved to a file for further use in the simulation process.